Update buttons to use new component approach to building
