+++ styleguide/templates/buttons.html
@@ -3,38 +3,38 @@
 <h1>Buttons</h1>
 <p>The standard button can be used in forms or for triggering javascript functions. You'll find it located in modals and forms.</p>
 <div class="rc_example">
-	{{ render_button('Button') | safe }}
+	{{ ui_element("buttons/button", {'text': 'Button'}) | safe }}
 </div>
 <h2>Button Styles</h2>
 <h3>Default buttons</h3>
 <div class="rc_example">
-	{{ render_button({'text': 'Default', 'color': 'default'}) | safe }}
-	{{ render_button({'text': 'Primary', 'color': 'primary'}) | safe }}
-	{{ render_button({'text': 'Danger', 'color': 'danger'}) | safe }}
-	{{ render_button({'text': 'Disabled', 'color': 'disabled'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Default', 'color': 'default'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Primary', 'color': 'primary'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Danger', 'color': 'danger'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Disabled', 'color': 'disabled'}) | safe }}
 </div>
 <h3>Link buttons</h3>
 <p>These buttons are styled to look like links, if you need to trigger a javascript function instead of a traditional hyperlink.</p>
 <div class="rc_example">
-	{{ render_button({'text': 'Default', 'color': 'default', 'style': 'link'}) | safe }}
-	{{ render_button({'text': 'Primary', 'color': 'primary', 'style': 'link'}) | safe }}
-	{{ render_button({'text': 'Danger', 'color': 'danger', 'style': 'link'}) | safe }}
-	{{ render_button({'text': 'Disabled', 'color': 'disabled', 'style': 'link'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Default', 'color': 'default', 'style': 'link'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Primary', 'color': 'primary', 'style': 'link'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Danger', 'color': 'danger', 'style': 'link'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Disabled', 'color': 'disabled', 'style': 'link'}) | safe }}
 </div>
 <h3>Block buttons</h3>
 <div class="rc_example">
-	{{ render_button({'text': 'Block style', 'color': 'default', 'style': 'block'}) | safe }}
-	{{ render_button({'text': 'Block style', 'color': 'primary', 'style': 'block'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Block style', 'color': 'default', 'style': 'block'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Block style', 'color': 'primary', 'style': 'block'}) | safe }}
 </div>
 <h3>Inline buttons</h3>
 <div class="rc_example">
-	{{ render_button({'text': 'Inline style', 'color': 'default', 'style': 'inline'}) | safe }}
-	{{ render_button({'text': 'Inline style', 'color': 'primary', 'style': 'inline'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Inline style', 'color': 'default', 'style': 'inline'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Inline style', 'color': 'primary', 'style': 'inline'}) | safe }}
 </div>
 <h2>Button sizes</h2>
 <div class="rc_example">
-	{{ render_button({'text': 'Small', 'size': 'small'}) | safe }}
-	{{ render_button({'text': 'Base', 'size': 'base'}) | safe }}
-	{{ render_button({'text': 'Large', 'size': 'large'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Small', 'size': 'small'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Base', 'size': 'base'}) | safe }}
+	{{ ui_element("buttons/button", {'text': 'Large', 'size': 'large'}) | safe }}
 </div>
 {% endblock %}
