+++ styleguide/templates/buttons/button.html
@@ -1,0 +1,1 @@
+<button{% if parameters['type'] %} type="{{ parameters['type'] }}"{% endif %}{% if parameters['name'] %} name="{{ parameters['name'] }}"{% endif %}{% if parameters['id'] %} id="{{ parameters['id'] }}"{% endif %}{% if parameters['data'] %}{% for label, value in parameters['data'].iteritems() %} data-{{ label }}="{{ value }}"{% endfor %}{% endif %}{% if parameters['onclick'] %} onClick="{{ parameters['onclick'] }}"{% endif %}{% if parameters['disabled'] %} disabled{% endif %} class="button{{ parameters['color'] }}{{ parameters['size'] }}{{ parameters['style'] }}">{{ parameters['text'] }}</button>
