--- styleguide/templates/components/buttons.html
@@ -1,1 +1,0 @@
-<button{{ attr_type | safe }}{{ attr_id | safe }}{{ data_target | safe }} class="{{ attr_class }}button{{ class_size }}{{ class_style }}{{ class_color }}"{{ attr_style | safe }}>{{ text }}{{ text_icon | safe}}</button>
