--- styleguide/views/__init__.py
@@ -1,2 +1,0 @@
-from buttons import *
-from text_inputs import *
